fix: improve strategy for hiding dynamic imports from bundlers
This is a slight recast of @Arash-11's fix in #117. The idea is that
since we need to postprocess the compiled output, we might as well
handle the entire process of hiding the dynamic output in post
processing. This substantially simplifies the postprocessing script. It
also eliminates the confusing behavior of relative paths with
importDynamic.

Co-authored-by: Arash-11 <arashnawy@gmail.com>