build: improve smoke test (#118)
* build: improve smoke test

Beef up the smoke test in scripts/build that checks the validity of the
compiles JavaScript so that it can detect the issue @Arash-11 is fixing
in #117.

* fix: improve strategy for hiding dynamic imports from bundlers

This is a slight recast of @Arash-11's fix in #117. The idea is that
since we need to postprocess the compiled output, we might as well
handle the entire process of hiding the dynamic output in post
processing. This substantially simplifies the postprocessing script. It
also eliminates the confusing behavior of relative paths with
importDynamic.

Co-authored-by: Arash-11 <arashnawy@gmail.com>

* include commonJS build in fix-bundle-ignore-imports.cjs

* Revert "include commonJS build in fix-bundle-ignore-imports.cjs"

This reverts commit b291d06fc64525455a100d79b25d286dd271df03.

---------

Co-authored-by: Arash-11 <arashnawy@gmail.com>