Fix bundling with old webpack versions (#127)
Old versions of webpack do not tolerate malformed keys in the `imports`
field of package.json. Keys in this map need to start with `#`, and we
had two entries that started with `@turbopuffer/turbopuffer`.

This commit both removes those malformed keys. I've verified with the
Stainless team that the malformed entries in the `imports` do not appear
to be necessary. They're likely to remove these entries upstream in the
future, but we can remove them now via custom code to get users
unblocked.

It also adds smoke tests for both old and new versions of webpack. The
test with webpack v5.83.0 and enhanced-resolve 5.14.1 does in fact fail
if the malformed import keys are left in place.

This should fix the bundling error that one of our users reported in the
community Slack with Next.js 14, which uses webpack v5.83 under the
hood.